btlp: the round's flow and the party's standing orders
Splits btlp_51 into src/btlp/roundflow.c and the head of btlp_61 into
talkorders.c and commandmenu.c, and names what they reach: BtlMarkMoveArea,
BtlReadyItemAction, BtlReadySpellAction, BtlCommandEntry, BtlOrdersMenu,
BtlConfigMenu, BtlRunTalkScene, and the two handlers that were sitting under
D_ names because they are code in a pointer table Ghidra never typed.

Matched: BtlStageOpen, BtlShowReadyMarkers, BtlPlaceMemberMarkers,
BtlPlaceFormation, BtlRunTalkScene, BtlTalkersStay, BtlTalkersJoin.

The three marker and formation routines were one defect three times over. Each
was written as a byte offset into the actor record with every field read
through it; the image indexes by slot and keeps at most one walking pointer,
for a field it writes more than once. Written the other way gcc strength
reduces a second field into a pointer of its own and the whole allocation
moves up a register. BtlPlaceFormation needed the table index adding up in the
image's order as well - the encounter's block first, then the character's pair.

Three levers came out of it and are in docs/Matching Nudges.md: a constant kept
in a local of its own stays out of the loop's preheader, which is the only way
to get it set up ahead of a pointer the source assigns; a two-armed dispatch on
0 and 1 has to be a switch, because an if chain inverts the first test and
drops a branch; and whether gcc lifts a constant out of a loop turns on where
its uses are, not how many.

Along the way: 0x8009DDD8 has no references anywhere in the image, so what
splat had as two symbols is one function; BtlActor's shadow of the chosen
action is named (move_kept, ail_line_kept, order_kept, targets_kept) and unk80
is targets, the mask of who the action hits; BtlShowMarker's prototype said
(slot, kind, level) where every call site says (slot, on, kind);
BtlTalkSceneStep returns the outcome rather than void; and the battle's copy of
the formation grid has a header of its own that folds onto the field's rather
than repeating its constants under new names.

BtlStageClose, BtlStageRound, BtlChooseEnemyMove and BtlTalkersLeaveField are
in the tree under NON_MATCHING with what is left written down; all eleven
targets still reproduce the ROM.

+++ src/btlp/soundbank.c
@@ -93,7 +93,7 @@ void BtlSoundClose(int slot)
 /* How the battle plays a sound: 44 call sites reach this, from the menu
    cursor to the spell effects. Stopping first is what makes retriggering a
    sound that is already playing restart it rather than do nothing. */
-void BtlSePlay(int slot, short seq)
+void BtlSePlay(int slot, int seq)
 {
     SsSepStop(g_btl_seq[slot], seq);
     SsSepPlay(g_btl_seq[slot], seq, SSPLAY_PLAY, 1);
